Reorganize python files
